remove ordered block id from device reduce by key
This seems to have had a negative effect in all cases: There is
a small overhead associated with fetching and updating the ordered
block id. By removing the ordered block id, the cases with 2 tiles
per block gain performance about equivalent to the speedup that
they provide compared to 1 tile per block. This also reduces
the amount of stuff that needs to be tuned a little.