run length encode: always use direct loading for value types
For run length encode, the value type is a counting iterator. Previously,
this algorithm always used block_load_transposed for the value loading
method, which stores the values in shared memory and then loads them later.
Because there is no actual loading involved with the counting iterator, this
loads to wasted instructions & shared memory usage.